Update travellingWaves.html
For the Kuramoto model, there is an likely an oversight between Theta (starting phase of each oscillator) and freq (natural frequency of each oscillator)

Please verify - the travelling waves you are seeing is due to the starting frequencies you set for each oscillator 2*(i/N). 

After changing this, you have to set kN to a higher value (about 7) to see travelling wave. Introducing.